Branch commit for highprecision

+++ FJ-VDMJ/src/main/java/org/overturetool/vdmj/definitions/CPUClassDefinition.java
@@ -116,7 +116,7 @@ public class CPUClassDefinition extends ClassDefinition
     		SeqValue opname = (SeqValue)octxt.lookup(varName("opname"));
     		NaturalValue priority = (NaturalValue)octxt.check(varName("priority"));
 
-    		cpu.setPriority(opname.stringValue(ctxt), priority.intValue(ctxt));
+    		cpu.setPriority(opname.stringValue(ctxt), priority.intValue(ctxt).longValue());
    			return new VoidValue();
 		}
 		catch (Exception e)
