+++ FJ-VDMJ/src/main/java/org/overturetool/vdmj/statements/PeriodicStatement.java
@@ -150,7 +150,7 @@ public class PeriodicStatement extends Statement
 			{
 				arg.location.hit();
 				argval = arg.eval(ctxt);
-				values[i] = argval.intValue(ctxt);
+				values[i] = argval.intValue(ctxt).longValue();
 
 				if (values[i] < 0)
 				{
