+++ FJ-VDMJ/src/main/java/org/overturetool/vdmj/values/CPUValue.java
@@ -49,7 +49,7 @@ public class CPUValue extends ObjectValue
 		SchedulingPolicy cpup = SchedulingPolicy.factory(parg.value.toUpperCase());
 		RealValue sarg = (RealValue)argvals.get(1);
 
-		resource = new CPUResource(cpup, sarg.value);
+		resource = new CPUResource(cpup, sarg.value.doubleValue());
 		deployed = new Vector<ObjectValue>();
 	}
