+++ FJ-VDMJ/src/main/java/org/overturetool/vdmj/values/TransactionValue.java
@@ -23,6 +23,8 @@
 
 package org.overturetool.vdmj.values;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.util.List;
 import java.util.ListIterator;
 import java.util.Vector;
@@ -202,25 +204,25 @@ public class TransactionValue extends UpdatableValue
 	}
 
 	@Override
-	public synchronized double realValue(Context ctxt) throws ValueException
+	public synchronized BigDecimal realValue(Context ctxt) throws ValueException
 	{
 		return select().realValue(ctxt);
 	}
 
 	@Override
-	public synchronized long intValue(Context ctxt) throws ValueException
+	public synchronized BigInteger intValue(Context ctxt) throws ValueException
 	{
 		return select().intValue(ctxt);
 	}
 
 	@Override
-	public synchronized long natValue(Context ctxt) throws ValueException
+	public synchronized BigInteger natValue(Context ctxt) throws ValueException
 	{
 		return select().nat1Value(ctxt);
 	}
 
 	@Override
-	public synchronized long nat1Value(Context ctxt) throws ValueException
+	public synchronized BigInteger nat1Value(Context ctxt) throws ValueException
 	{
 		return select().nat1Value(ctxt);
 	}
