+++ FJ-VDMJ/src/test/java/junit/overture/OvertureTest.java
@@ -366,7 +366,7 @@ abstract public class OvertureTest extends TestCase
 
 			for (Value ex: assertions.seqValue(null))
 			{
-				long n = ex.intValue(null);
+				long n = ex.intValue(null).longValue();
 				expected.add(n);
 			}
