+++ FJ-VDMJUnit/src/main/java/vdmjunit/VDMJUnitTest.java
@@ -23,7 +23,10 @@
 
 package vdmjunit;
 
+import java.math.BigDecimal;
+import java.math.BigInteger;
 import java.nio.charset.Charset;
+
 import static org.junit.Assert.fail;
 
 import org.overturetool.vdmj.Release;
@@ -207,7 +210,7 @@ public abstract class VDMJUnitTest
 	 * 			or the specification has not been initialized, or if the result is not
 	 * 			a VDM integer value.
 	 */
-	protected long runInt(String expression) throws Exception
+	protected BigInteger runInt(String expression) throws Exception
 	{
 		return run(expression).intValue(null);
 	}
@@ -222,7 +225,7 @@ public abstract class VDMJUnitTest
 	 * 			or the specification has not been initialized, or if the result is not
 	 * 			a VDM real value.
 	 */
-	protected double runReal(String expression) throws Exception
+	protected BigDecimal runReal(String expression) throws Exception
 	{
 		return run(expression).realValue(null);
 	}
